+++ src/zugbruecke/wenv.py
@@ -26,7 +26,6 @@ specific language governing rights and limitations under the License.
 
 """
 
-
 # +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 # IMPORT
 # +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
@@ -40,27 +39,27 @@ import urllib.request
 import zipfile
 
 from .core.config import config_class
+from .core.log import c
 
 
 # +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 # CONST
 # +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
-HELP_STR = """wenv - the Wine Python environment
-
-- wenv init: sets up an environment (including Python interpreter, pip and pytest)
-- wenv help: prints this help text
-- wenv python: the Python interpreter
-
-Beyond that, the following scripts and modules are installed and available:
-
-{scripts:s}
-"""
-
 COVERAGE_STARTUP = """
 import coverage
 coverage.process_startup()
 """
+
+HELP_STR = """wenv - the Wine Python environment
+
+{CLIS:s}
+
+The following interpreters, scripts and modules are installed and available:
+
+{SCRIPTS:s}
+"""
+
 
 # +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 # WINE-PYTHON ENVIRONMENT CLASS
@@ -77,67 +76,43 @@ class env_class:
 
 		# Get config
 		if parameter is None:
-			self.p = config_class()
+			self._p = config_class()
 		else:
 			if not isinstance(parameter, dict):
 				raise TypeError('parameter is not a dictionary')
-			self.p = parameter
+			self._p = parameter
 
-		# Get Python environment paths
-		self._path_dict_ = env_class.__get_wine_python_paths__(
-			self.p['pythonprefix'], self.p['version']
-			)
-		# Get Python commands and scripts
-		self._cmd_dict_ = env_class.__get_command_dict__(
-			self._path_dict_['interpreter'], self._path_dict_['scripts']
-			)
-		# Get environment variables
-		self._envvar_dict_ = env_class.__get_environment_variables__(
-			self.p['wineprefix'], self.p['winedebug'], self.p['arch']
-			)
-		# Get Wine cmd names
-		self._wine_dict_ = {'win32': 'wine', 'win64': 'wine64'}
+		# Init internal dicts
+		self._init_dicts()
 
 
-	@staticmethod
-	def __get_environment_variables__(wineprefix, winedebug, arch):
+	def _init_dicts(self):
 
-		return dict(
-			WINEARCH = arch, # Architecture
-			WINEPREFIX = wineprefix, # Wine prefix / directory
-			WINEDLLOVERRIDES = 'mscoree=d', # Disable MONO: https://unix.stackexchange.com/a/191609
-			WINEDEBUG = winedebug, # Wine debug level
-			)
+		# Init Wine cmd names
+		self._wine_dict = {'win32': 'wine', 'win64': 'wine64'}
 
-
-	@staticmethod
-	def __get_command_dict__(interpreter_path, scripts_path):
-
-		out = {'python': interpreter_path}
-
-		if not os.path.exists(scripts_path):
-			return out
-
-		scripts = os.listdir(scripts_path)
-		for script in scripts:
-			if not script.lower().endswith('.exe'):
-				continue
-			out[script[:-4]] = os.path.join(scripts_path, script)
-
-		return out
+		# Init Python environment paths
+		self._init_path_dict()
+		# Init Python commands and scripts
+		self._init_cmd_dict()
+		# Init internal CLI commands
+		self._init_cli_dict()
+		# Init environment variables
+		self._init_envvar_dict()
 
 
-	@staticmethod
-	def __get_wine_python_paths__(pythonprefix, version):
+	def _init_path_dict(self):
+
+		version_string = ''.join(self._p['version'].split('.')[0:2])
 
 		# python standard library
-		lib_path = os.path.join(pythonprefix, 'Lib')
+		lib_path = os.path.join(self._p['pythonprefix'], 'Lib')
 		# site-packages
 		sitepackages_path = os.path.join(lib_path, 'site-packages')
 		# python interpreter
-		interpreter_path = os.path.join(pythonprefix, 'python.exe')
+		interpreter_path = os.path.join(self._p['pythonprefix'], 'python.exe')
 		# scripts
-		scripts_path = os.path.join(pythonprefix, 'Scripts')
+		scripts_path = os.path.join(self._p['pythonprefix'], 'Scripts')
 		# pip
 		pip_path = os.path.join(scripts_path, 'pip.exe')
 		# pytest
@@ -145,13 +120,11 @@ class env_class:
 		# coverage
 		coverage_path = os.path.join(scripts_path, 'coverage.exe')
 		# stdlib zip filename
-		stdlibzip_path = os.path.join(
-			pythonprefix,
-			'python%s%s.zip' % (version.split('.')[0], version.split('.')[1])
-			)
+		stdlibzip_path = os.path.join(self._p['pythonprefix'], 'python%s.zip' % version_string)
+		# pth filename (library path)
+		pth_path = os.path.join(self._p['pythonprefix'], 'python%s._pth' % version_string)
 
-		# Return dict
-		return dict(
+		self._path_dict = dict(
 			lib = lib_path,
 			sitepackages = sitepackages_path,
 			scripts = scripts_path,
@@ -160,6 +133,44 @@ class env_class:
 			pytest = pytest_path,
 			coverage = coverage_path,
 			stdlibzip = stdlibzip_path,
+			pth = pth_path,
+			)
+
+
+	def _init_cmd_dict(self):
+
+		out = {'python': self._path_dict['interpreter']} # TODO check!
+
+		if not os.path.exists(self._path_dict['scripts']):
+			self._cmd_dict = out
+			return
+
+		scripts = os.listdir(self._path_dict['scripts'])
+		for script in scripts:
+			if not script.lower().endswith('.exe'):
+				continue
+			out[script[:-4]] = os.path.join(self._path_dict['scripts'], script)
+
+		self._cmd_dict = out
+
+
+	def _init_cli_dict(self):
+
+		self._cli_dict = {
+			item[5:]: getattr(self, item)
+			for item in dir(self)
+			if item.startswith('_cli_') and hasattr(getattr(self, item), '__call__')
+			}
+
+
+	def _init_envvar_dict(self):
+
+		self._envvar_dict = dict(
+			WINEARCH = self._p['arch'], # Architecture
+			WINEPREFIX = self._p['wineprefix'], # Wine prefix / directory
+			WINEDLLOVERRIDES = 'mscoree=d', # Disable MONO: https://unix.stackexchange.com/a/191609
+			WINEDEBUG = self._p['winedebug'], # Wine debug level
+			PYTHONHOME = self._p['pythonprefix'], # Python home for Wine Python (can be a Unix path)
 			)
 
 
@@ -173,15 +184,15 @@ class env_class:
 			raise TypeError('overwrite is not a boolean')
 
 		# Does it exist?
-		if os.path.exists(self.p['wineprefix']):
+		if os.path.exists(self._p['wineprefix']):
 			# Exit if overwrite flag is not set
 			if not overwrite:
 				return
 			# Delete if overwrite is set
-			shutil.rmtree(self.p['wineprefix'])
+			shutil.rmtree(self._p['wineprefix'])
 
 		# Start wine server into prepared environment
-		subprocess.Popen(['wineboot', '-i']).wait()
+		subprocess.Popen(['wineboot', '-i'], env = self._envvar_dict).wait()
 
 
 	def setup_python(self, overwrite = False):
@@ -191,24 +202,24 @@ class env_class:
 
 		# File name for python stand-alone zip file
 		pyarchive = 'python-%s-embed-%s.zip' % (
-			self.p['version'],
-			'amd64' if self.p['arch'] == 'win64' else self.p['arch']
+			self._p['version'],
+			'amd64' if self._p['arch'] == 'win64' else self._p['arch']
 			)
 		# Compute full URL of Python stand-alone zip file
-		pyurl = 'https://www.python.org/ftp/python/%s/%s' % (self.p['version'], pyarchive)
+		pyurl = 'https://www.python.org/ftp/python/%s/%s' % (self._p['version'], pyarchive)
 
 		# Is there a pre-existing Python installation with identical parameters?
-		preexisting = os.path.isfile(self._path_dict_['interpreter'])
+		preexisting = os.path.isfile(self._path_dict['interpreter'])
 
 		# Is there a preexisting installation and should it be overwritten?
 		if preexisting and overwrite:
 			# Delete folder
-			shutil.rmtree(self.p['pythonprefix'])
+			shutil.rmtree(self._p['pythonprefix'])
 
 		# Make sure the target directory exists
-		if not os.path.exists(self.p['pythonprefix']):
+		if not os.path.exists(self._p['pythonprefix']):
 			# Create folder
-			os.makedirs(self.p['pythonprefix'])
+			os.makedirs(self._p['pythonprefix'])
 
 		# Only do if Python is not there OR if should be overwritten
 		if overwrite or not preexisting:
@@ -220,21 +231,24 @@ class env_class:
 				archive_zip.write(u.read())
 			# Unpack from memory to disk
 			with zipfile.ZipFile(archive_zip) as f:
-				f.extractall(path = self.p['pythonprefix']) # Directory created if required
+				f.extractall(path = self._p['pythonprefix']) # Directory created if required
 
 			# Unpack Python library from embedded zip on disk
-			with zipfile.ZipFile(self._path_dict_['stdlibzip'], 'r') as f:
-				f.extractall(path = self._path_dict_['lib']) # Directory created if required
+			with zipfile.ZipFile(self._path_dict['stdlibzip'], 'r') as f:
+				f.extractall(path = self._path_dict['lib']) # Directory created if required
 			# Remove Python library zip from disk
-			os.remove(self._path_dict_['stdlibzip'])
+			os.remove(self._path_dict['stdlibzip'])
+
+			# HACK: Fix library path in pth-file (CPython >= 3.6)
+			os.unlink(self._path_dict['pth'])
 
 		# Create site-packages folder if it does not exist
-		if not os.path.exists(self._path_dict_['sitepackages']):
+		if not os.path.exists(self._path_dict['sitepackages']):
 			# Create folder
-			os.makedirs(self._path_dict_['sitepackages'])
+			os.makedirs(self._path_dict['sitepackages'])
 
 		# Package path in wine-python site-packages
-		wine_pkg_path = os.path.abspath(os.path.join(self._path_dict_['sitepackages'], 'zugbruecke'))
+		wine_pkg_path = os.path.abspath(os.path.join(self._path_dict['sitepackages'], 'zugbruecke'))
 
 		# Package path in unix-python site-packages
 		unix_pkg_path = os.path.abspath(os.path.dirname(__file__))
@@ -245,10 +259,47 @@ class env_class:
 			os.symlink(unix_pkg_path, wine_pkg_path)
 
 
+	def wine_47766_workaround(self):
+		"""
+		PathAllocCanonicalize treats path segments start with dots wrong.
+		https://bugs.winehq.org/show_bug.cgi?id=47766
+		"""
+
+		is_clean = lambda path: not any([seg.startswith('.') for seg in path.split(os.path.sep)])
+
+		pythonprefix = os.path.abspath(self._p['pythonprefix'])
+
+		if pythonprefix != self._p['pythonprefix']:
+			self._p['pythonprefix'] = pythonprefix
+			self._init_dicts()
+
+		if is_clean(self._p['pythonprefix']):
+			return
+
+		import tempfile, hashlib
+		link_path = os.path.join(
+			tempfile.gettempdir(),
+			'wenv-' + hashlib.sha256(self._p['pythonprefix'].encode('utf-8')).hexdigest()[:8],
+			)
+		if not is_clean(link_path):
+			raise OSError('unable to create clean link path: "{LINK:s}"'.format(LINK = link_path))
+		if not os.path.exists(link_path):
+			os.symlink(self._p['pythonprefix'], link_path)
+		if not os.path.exists(link_path):
+			raise OSError('"{LINK:s}" could not be created'.format(LINK = link_path))
+		if not os.path.islink(link_path):
+			raise OSError('"{LINK:s}" is not a symlink'.format(LINK = link_path))
+		if os.readlink(link_path) != pythonprefix:
+			raise OSError('"{LINK:s}" points to the wrong source'.format(LINK = link_path))
+
+		self._p['pythonprefix'] = link_path
+		self._init_dicts()
+
+
 	def setup_pip(self):
 
 		# Exit if it exists
-		if os.path.isfile(self._path_dict_['pip']):
+		if os.path.isfile(self._path_dict['pip']):
 			return
 
 		# Download get-pip.py into memory
@@ -265,7 +316,7 @@ class env_class:
 	def setup_pytest(self):
 
 		# Exit if it exists
-		if os.path.isfile(self._path_dict_['pytest']):
+		if os.path.isfile(self._path_dict['pytest']):
 			return
 
 		# Run pip install
@@ -275,119 +326,119 @@ class env_class:
 	def setup_coverage(self):
 
 		# Exit if it exists
-		if os.path.isfile(self._path_dict_['coverage']):
+		if os.path.isfile(self._path_dict['coverage']):
 			return
 
 		# Run pip install
 		subprocess.Popen(['wenv', 'pip', 'install', 'coverage']).wait()
 
 		# Ensure that coverage is started with the Python interpreter
-		siteconfig_path = os.path.join(self._path_dict_['sitepackages'], 'sitecustomize.py')
+		siteconfig_path = os.path.join(self._path_dict['sitepackages'], 'sitecustomize.py')
+		siteconfig_cnt = ''
 		if os.path.isfile(siteconfig_path):
 			with open(siteconfig_path, 'r') as f:
-				siteconfig_cnt = f.read(COVERAGE_STARTUP)
+				siteconfig_cnt = f.read()
 			if COVERAGE_STARTUP in siteconfig_cnt:
 				return
 		with open(siteconfig_path, 'w') as f:
-			f.write(COVERAGE_STARTUP)
+			f.write(siteconfig_cnt + '\n' + COVERAGE_STARTUP)
 
 
 # +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-# ROUTINES
+# CLI
 # +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
-	def set_environment_variables(self):
+	def _cli_init(self):
+		"sets up an environment (including Python interpreter, pip and pytest)"
 
-		for k, v in self._envvar_dict_.items():
-			os.environ[k] = v
-
-
-# +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-# CLI HELPER
-# +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-
-	def __cmd_init__(self):
-		self.set_environment_variables()
 		self.setup_prefix()
 		self.setup_python()
+		self.wine_47766_workaround() # must run after setup_python and before setup_pip
 		self.setup_pip()
 		self.setup_pytest()
 		self.setup_coverage()
 
 
-	def __cmd_help__(self):
-		sys.stdout.write(HELP_STR.format(scripts = '\n'.join([
-			'- wenv {script:s}'.format(script = key)
-			for key in sorted(self._cmd_dict_.keys())
-			if key != 'python'
-			])))
+	def _cli_help(self):
+		"prints this help text"
+
+		def colorize(text):
+			for color, key in (
+				('GREEN', 'python'),
+				('CYAN', 'pip'),
+				('YELLOW', 'wheel'),
+				('MAGENTA', 'pytest'),
+				('MAGENTA', 'py.test'),
+				):
+				text = text.replace(key, c[color] + key + c['RESET'])
+			return text
+
+		sys.stdout.write(HELP_STR.format(
+			CLIS = '\n'.join([
+				'- wenv {CLI:s}: {HELP:s}'.format(
+					CLI = key,
+					HELP = self._cli_dict[key].__doc__,
+					)
+				for key in sorted(self._cli_dict.keys())
+				]),
+			SCRIPTS = colorize('\n'.join([
+				'- wenv {SCRIPT:s}'.format(
+					SCRIPT = key,
+					)
+				for key in sorted(self._cmd_dict.keys())
+				]))
+			))
 		sys.stdout.flush()
 
 
 	def cli(self):
 
-		def prefix(data, options = False):
-			return '___%d___%s' % ((1 if options else 0), data)
-		def push_return(data):
-			sys.stdout.write(data)
-			sys.stdout.flush()
-			sys.exit()
-
-		# No command was passed?
+		# No command passed
 		if len(sys.argv) < 2:
-			# Exit: Nothing to do
-			return
+			sys.stderr.write('There was no command passed.\n')
+			sys.stderr.flush()
+			sys.exit(1)
 
-		# Not (yet) a configuration call
-		configure = False
-		# Are there one or more arguments / options?
-		if len(sys.argv) >= 3:
-			# Is this a configuration call?
-			if sys.argv[2] == '--configure':
-				# Re-set flag
-				configure = True
+		# Separate command and arguments
+		cmd, param = sys.argv[1], sys.argv[2:]
 
-		# Get command
-		cmd = sys.argv[1]
+		# Special CLI command
+		if cmd in self._cli_dict.keys():
+			self._cli_dict[cmd]()
+			sys.exit(0)
 
-		# Special commands
-		special_list = ['init', 'help']
-		# Special command, first stage?
-		if configure and (cmd in special_list):
-			# Push new Python command for "interactive" second stage
-			push_return(prefix(
-				'python3 -m zugbruecke.wenv %s' % cmd, options = False
-				))
-		# Special command, second stage?
-		if not configure and (cmd in special_list):
-			# Run matching routine
-			getattr(self, '__cmd_%s__' % cmd)()
-			# Nothing more to do, leave
-			return
+		# Command is unknown
+		if cmd not in self._cmd_dict.keys():
+			sys.stderr.write('Unknown command or script: "{CMD:s}"\n'.format(CMD = cmd))
+			sys.stderr.flush()
+			sys.exit(1)
 
-		# If the command is unknown ...
-		if configure and (cmd not in self._cmd_dict_.keys()):
-			push_return(prefix('echo Unknown command or script: "%s"' % cmd, options = False))
+		# Get Wine depending on arch
+		wine = self._wine_dict[self._p['arch']]
 
-		# Nothing more to do
-		if not configure:
-			return
+		self.wine_47766_workaround()
 
-		# Prepare environment variables for shell export
-		cmd_list = ['%s=%s' % (k, v) for k, v in self._envvar_dict_.items()]
-		# Append actual "interactive" command
-		cmd_list.append(prefix(
-			'%s "%s"' % (self._wine_dict_[self.p['arch']], self._cmd_dict_[cmd]),
-			options = True,
-			))
-		# Push command stack and leave
-		push_return('#'.join(cmd_list))
+		# Replace this process with Wine
+		os.execvpe(
+			wine,
+			(wine, self._cmd_dict[cmd], *param),
+			self._envvar_dict,
+			)
 
 
 # +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
-# MAIN
+# CLI EXPORT
+# +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
+
+def cli():
+
+	env_class().cli()
+
+
+# +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
+# MODULE ENTRY POINT
 # +++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++++
 
 if __name__ == '__main__':
 
-	env_class().cli()
+	cli()